[Event Hubs] move receiveHandler into own file
